requirements.txt
numpy>=1.20.3
pandas>=1.3.4
Pillow==9.1.0
requests==2.26.0
scikit_learn==1.0.2
streamlit==1.8.1
